Refactor middleware and validation methods into separate files
