+++ lib/expose.js
@@ -2,46 +2,8 @@ const Router = require('koa-router');
 
 const defaults = require('./defaults');
 const handlers = require('./handlers');
-const response = require('./response');
-
-async function initSwatchCtx(koaCtx, next) {
-  // Create the swatchCtx object on the KOA ctx
-  koaCtx.swatchCtx = {};
-
-  // Continue chain of handlers
-  await next();
-}
-
-function wrapMiddleware(fn) {
-  async function middlewareWrapper(koaCtx, next) {
-    try {
-      await fn(koaCtx, next);
-    } catch (error) {
-      // Auth error should trigger failure response
-      response.errorResponse(koaCtx, error);
-    }
-  }
-
-  return middlewareWrapper;
-}
-
-function validationMiddleware(validateFn, verb) {
-  async function validationWrapper(koaCtx, next) {
-    try {
-      // Get the params from request body and validate
-      //  Updates koaCtx.swatchCtx with validated params
-      const requestParams = handlers[verb](koaCtx);
-      validateFn(koaCtx, requestParams);
-
-      await next();
-    } catch (error) {
-      // Validation error should trigger failure response
-      response.errorResponse(koaCtx, error);
-    }
-  }
-
-  return validationWrapper;
-}
+const middleware = require('./middleware');
+const validation = require('./validation');
 
 function expose(app, methods, opts) {
   const options = defaults(opts);
@@ -57,17 +19,17 @@ function expose(app, methods, opts) {
       const methodMiddleware = method.metadata.middleware;
 
       const adapter = noAuth ?
-        [initSwatchCtx] :
-        [initSwatchCtx, options.authAdapter];
+        [middleware.initSwatchCtx] :
+        [middleware.initSwatchCtx, options.authAdapter];
       const validator = adapter.concat(
-        [validationMiddleware(method.validate, verb)],
+        [validation.wrapMiddleware(method.validate, verb)],
       );
-      const middleware = validator.concat(
-        methodMiddleware.map(wrapMiddleware),
+      const middlewareArray = validator.concat(
+        methodMiddleware.map(middleware.wrapMiddleware),
       );
 
       const handler = handlers.handler(method.handle);
-      const handlerList = middleware.concat([handler]);
+      const handlerList = middlewareArray.concat([handler]);
 
       router[verb](path, ...handlerList);
     }
